python/requirements.txt
json
requests
qrcode[pil]
python-dateutil
python-dotenv